Changed shell logic to be OS specific + updates

+++ lua/config/options.lua
@@ -7,10 +7,14 @@ vim.o.expandtab = true -- Pressing the TAB key will insert spaces instead of a T
 vim.o.softtabstop = 4 -- Number of spaces inserted instead of a TAB character
 vim.o.shiftwidth = 4 -- Number of spaces inserted when indenting
 
-vim.cmd([[
-let &shell = executable('pwsh') ? 'pwsh' : 'powershell'
-let &shellcmdflag = '-NoLogo -ExecutionPolicy RemoteSigned -Command [Console]::InputEncoding=[Console]::OutputEncoding=[System.Text.UTF8Encoding]::new();$PSDefaultParameterValues[''Out-File:Encoding'']=''utf8'';Remove-Alias -Force -ErrorAction SilentlyContinue tee;'
-let &shellredir = '2>&1 | %%{ "$_" } | Out-File %s; exit $LastExitCode'
-let &shellpipe  = '2>&1 | %%{ "$_" } | tee %s; exit $LastExitCode'
-set shellquote= shellxquote=
+-- Only overwrite shell on Windows, leave default on other OSes
+if vim.fn.has("win32") == 1 then
+  -- Windows, use powershell instead of cmd
+  vim.cmd([[
+    let &shell = executable('pwsh') ? 'pwsh' : 'powershell'
+    let &shellcmdflag = '-NoLogo -ExecutionPolicy RemoteSigned -Command [Console]::InputEncoding=[Console]::OutputEncoding=[System.Text.UTF8Encoding]::new();$PSDefaultParameterValues[''Out-File:Encoding'']=''utf8'';Remove-Alias -Force -ErrorAction SilentlyContinue tee;'
+    let &shellredir = '2>&1 | %%{ "$_" } | Out-File %s; exit $LastExitCode'
+    let &shellpipe  = '2>&1 | %%{ "$_" } | tee %s; exit $LastExitCode'
+    set shellquote= shellxquote=
 ]])
+end
